change len type to size_t

+++ include/libssh2.h
@@ -852,9 +852,9 @@ LIBSSH2_API int libssh2_channel_x11_req_ex(LIBSSH2_CHANNEL *channel,
 
 LIBSSH2_API int libssh2_channel_signal_ex(LIBSSH2_CHANNEL *channel,
                                           const char *signame,
-                                          unsigned int signame_len);
+                                          size_t signame_len);
 #define libssh2_channel_signal(channel, signame) \
-    libssh2_channel_signal_ex((channel), signame, (unsigned int)strlen(signame))
+    libssh2_channel_signal_ex((channel), signame, strlen(signame))
 
 LIBSSH2_API int libssh2_channel_process_startup(LIBSSH2_CHANNEL *channel,
                                                 const char *request,
